Release v0.1.7: Code quality improvements with Black formatting and strict MyPy type checking
🧹 Code Quality Improvements:
- Applied Black formatting to all 7 source files for consistent code style
- Added comprehensive MyPy type annotations with strict mode compliance
- Installed external type stubs (pandas-stubs, types-psutil)
- Enhanced developer experience with complete type hints

✅ All Quality Checks Pass:
- Black formatting: ✅ All files consistent
- MyPy strict mode: ✅ Zero type errors across 7 files
- Functionality: ✅ Core features working correctly

🔧 Technical Improvements:
- Fixed generic type parameters (dict → Dict[str, Any])
- Added missing return type annotations
- Corrected complex nested type annotations
- Improved IDE support and autocompletion
- Enhanced code maintainability and reliability

+++ xraylabtool/constants.py
@@ -309,7 +309,7 @@ def critical_angle_degrees(dispersion: float) -> float:
         raise ValueError("Dispersion coefficient must be positive")
 
     theta_c_rad = np.sqrt(2.0 * dispersion)
-    return theta_c_rad * RADIANS_TO_DEGREES
+    return float(theta_c_rad * RADIANS_TO_DEGREES)
 
 
 def attenuation_length_cm(wavelength_angstrom: float, absorption: float) -> float:
